Data Cleaning using MS Excel
The spreadsheet consists of a dirty data cleaned and transformed into a usable format using MS Excel. The dirty data and the cleaned data are both in the spreadsheet. I have documented the step-by-step walk-through of how I cleaned the data from scratch in the 'Report on how the data was cleaned.docx' word document. The final data can be used without any mismatched datatypes or crossing-over of values into different cells.